libzmqutil: create new zmqutil internal library
Problem: We would like to remove the libzmq dependency on libflux-core.
Doing so requires us to put libzmq dependent functions somewhere else.

Solution: Create a new internal library libzmqutil and put
flux_msg_sendzsock() and flux_msg_recvzsock() in there.  Rename functions
to zmqutil_msg_send() and zmqutil_msg_recv() respectively.  Move related
unit tests as well.  Update all callers, include new headers,
and link to new internal library.
