+++ configure.ac
@@ -502,6 +502,7 @@ AC_CONFIG_FILES( \
   src/common/librlist/Makefile \
   src/common/libczmqcontainers/Makefile \
   src/common/libccan/Makefile \
+  src/common/libzmqutil/Makefile \
   src/bindings/Makefile \
   src/bindings/lua/Makefile \
   src/bindings/python/Makefile \
