+++ src/broker/overlay.c
@@ -19,6 +19,7 @@
 #include <jansson.h>
 #include <uuid.h>
 
+#include "src/common/libzmqutil/msg_zsock.h"
 #include "src/common/libczmqcontainers/czmq_containers.h"
 #include "src/common/libutil/log.h"
 #include "src/common/libutil/kary.h"
@@ -349,7 +350,7 @@ static int overlay_sendmsg_parent (struct overlay *ov, const flux_msg_t *msg)
         errno = EHOSTUNREACH;
         goto done;
     }
-    rc = flux_msg_sendzsock (ov->parent.zsock, msg);
+    rc = zmqutil_msg_send (ov->parent.zsock, msg);
     if (rc == 0)
         ov->parent.lastsent = flux_reactor_now (ov->reactor);
 done:
@@ -501,7 +502,7 @@ static int overlay_sendmsg_child (struct overlay *ov, const flux_msg_t *msg)
         errno = EHOSTUNREACH;
         goto done;
     }
-    rc = flux_msg_sendzsock_ex (ov->bind_zsock, msg, true);
+    rc = zmqutil_msg_send_ex (ov->bind_zsock, msg, true);
 done:
     return rc;
 }
@@ -569,7 +570,7 @@ static void child_cb (flux_reactor_t *r, flux_watcher_t *w,
     struct child *child;
     int status;
 
-    if (!(msg = flux_msg_recvzsock (ov->bind_zsock)))
+    if (!(msg = zmqutil_msg_recv (ov->bind_zsock)))
         return;
     if (flux_msg_get_type (msg, &type) < 0
         || !(sender = flux_msg_route_last (msg)))
@@ -630,7 +631,7 @@ static void parent_cb (flux_reactor_t *r, flux_watcher_t *w,
     int type;
     const char *topic = NULL;
 
-    if (!(msg = flux_msg_recvzsock (ov->parent.zsock)))
+    if (!(msg = zmqutil_msg_recv (ov->parent.zsock)))
         return;
     if (flux_msg_get_type (msg, &type) < 0) {
         goto drop;
