+++ src/common/Makefile.am
@@ -23,7 +23,8 @@ SUBDIRS = \
 	libhostlist \
 	librlist \
 	libczmqcontainers \
-	libccan
+	libccan \
+	libzmqutil
 
 AM_CFLAGS = $(WARNING_CFLAGS) $(CODE_COVERAGE_CFLAGS)
 AM_LDFLAGS = $(CODE_COVERAGE_LIBS)
