+++ src/common/libflux/test/message.c
@@ -12,10 +12,11 @@
 #include "config.h"
 #endif
 #include <stdbool.h>
-#include <czmq.h>
 #include <errno.h>
 #include <stdio.h>
 #include <jansson.h>
+#include <string.h>
+#include <assert.h>
 
 #include "src/common/libflux/message.h"
 #include "src/common/libtap/tap.h"
@@ -949,81 +950,6 @@ void check_encode (void)
     flux_msg_destroy (msg2);
 }
 
-void check_sendzsock (void)
-{
-    zsock_t *zsock[2] = { NULL, NULL };
-    flux_msg_t *any, *msg, *msg2;
-    const char *topic;
-    int type;
-    const char *uri = "inproc://test";
-
-    /* zsys boiler plate:
-     * appears to be needed to avoid atexit assertions when lives_ok()
-     * macro (which calls fork()) is used.
-     */
-    zsys_init ();
-    zsys_set_logstream (stderr);
-    zsys_set_logident ("test_message.t");
-    zsys_handler_set (NULL);
-    zsys_set_linger (5); // msec
-
-    ok ((zsock[0] = zsock_new_pair (NULL)) != NULL
-                    && zsock_bind (zsock[0], "%s", uri) == 0
-                    && (zsock[1] = zsock_new_pair (uri)) != NULL,
-        "got inproc socket pair");
-
-    if (!(any = flux_msg_create (FLUX_MSGTYPE_ANY)))
-        BAIL_OUT ("flux_msg_create failed");
-
-    ok ((msg = flux_msg_create (FLUX_MSGTYPE_REQUEST)) != NULL
-            && flux_msg_set_topic (msg, "foo.bar") == 0,
-        "created test message");
-
-    /* corner case tests */
-    ok (flux_msg_sendzsock (NULL, msg) < 0 && errno == EINVAL,
-        "flux_msg_sendzsock returns < 0 and EINVAL on dest = NULL");
-    ok (flux_msg_sendzsock (zsock[1], any) < 0 && errno == EPROTO,
-        "flux_msg_sendzsock returns < 0 and EPROTO on msg w/ type = ANY");
-    ok (flux_msg_sendzsock_ex (NULL, msg, true) < 0 && errno == EINVAL,
-        "flux_msg_sendzsock_ex returns < 0 and EINVAL on dest = NULL");
-    ok (flux_msg_sendzsock_ex (zsock[1], any, true) < 0 && errno == EPROTO,
-        "flux_msg_sendzsock_ex returns < 0 and EPROTO on msg w/ type = ANY");
-    ok (flux_msg_recvzsock (NULL) == NULL && errno == EINVAL,
-        "flux_msg_recvzsock returns NULL and EINVAL on dest = NULL");
-
-    ok (flux_msg_sendzsock (zsock[1], msg) == 0,
-        "flux_msg_sendzsock works");
-    ok ((msg2 = flux_msg_recvzsock (zsock[0])) != NULL,
-        "flux_msg_recvzsock works");
-    ok (flux_msg_get_type (msg2, &type) == 0 && type == FLUX_MSGTYPE_REQUEST
-            && flux_msg_get_topic (msg2, &topic) == 0
-            && !strcmp (topic, "foo.bar")
-            && flux_msg_has_payload (msg2) == false,
-        "decoded message looks like what was sent");
-    flux_msg_destroy (msg2);
-
-    /* Send it again.
-     */
-    ok (flux_msg_sendzsock (zsock[1], msg) == 0,
-        "try2: flux_msg_sendzsock works");
-    ok ((msg2 = flux_msg_recvzsock (zsock[0])) != NULL,
-        "try2: flux_msg_recvzsock works");
-    ok (flux_msg_get_type (msg2, &type) == 0 && type == FLUX_MSGTYPE_REQUEST
-            && flux_msg_get_topic (msg2, &topic) == 0
-            && !strcmp (topic, "foo.bar")
-            && flux_msg_has_payload (msg2) == false,
-        "try2: decoded message looks like what was sent");
-    flux_msg_destroy (msg2);
-    flux_msg_destroy (msg);
-
-    zsock_destroy (&zsock[0]);
-    zsock_destroy (&zsock[1]);
-
-    /* zsys boiler plate - see note above
-     */
-    zsys_shutdown();
-}
-
 void *myfree_arg = NULL;
 void myfree (void *arg)
 {
@@ -1320,7 +1246,6 @@ int main (int argc, char *argv[])
     check_cmp ();
 
     check_encode ();
-    check_sendzsock ();
 
     check_refcount();
