+++ src/common/libterminus/Makefile.am
@@ -36,6 +36,7 @@ test_ldadd = \
 	$(top_builddir)/src/common/libterminus/libterminus.la \
 	$(top_builddir)/src/common/libutil/libutil.la \
 	$(top_builddir)/src/common/libtestutil/libtestutil.la \
+	$(top_builddir)/src/common/libzmqutil/libzmqutil.la \
         $(top_builddir)/src/common/libflux-internal.la \
         $(top_builddir)/src/common/libflux-core.la \
         $(top_builddir)/src/common/libtap/libtap.la
